Simplify dropdown openers and CSS
Since we're now closing the dropdown after a click, it doesn't matter that the text changes and would move the dropdown. This significantly simplifies the styling and HTML for the opener, yay! Would be nice to be able to do this with the filter dropdown as well, but since it remains open as the text changes this may not be possible.

The tab buttons in the filter menu can be controlled by their aria-pressed attribute.

No need to manually call _toggleFilterDropdown, just use auto-open.

Also some other minor stuff.