Remove link role from course tile
This was causing some issues with screen readers, and it's not actually true. For a sighted user, it may seem like the entire d2l-course-tile is a link, but in reality it's actually two things - a link (the text/image) to the course, and a menu. This means that the whole tile shouldn't have a link ARIA role.